// src/types/types.ts
export interface Person {
  url: string;
  name: string;
  gender: string;
  birth_year: string;
}

export interface PersonDetails {
  url: string;
  name: string;
  gender: string;
  birth_year: string;
  height: number;
  mass: number;
  hair_color: string;
  skin_color: string;
  eye_color: string;
}